Fix username page when visit. Separate to MyMessages.
